# feat: rewrite to TS + Major performance boosts