--- a/.github/workflows/codequality.yml
+++ b/.github/workflows/codequality.yml
@@ -1,0 +1,56 @@
+name: Code Quality
+
+on:
+    push:
+        branches:
+            - master
+    pull_request:
+
+jobs:
+    ESLint:
+        name: ESLint
+        runs-on: ubuntu-latest
+        steps:
+            - name: Checkout Project
+              uses: actions/checkout@v1
+            - name: Use Node.js 12
+              uses: actions/setup-node@v1
+              with:
+                  node-version: 12
+            - name: Restore CI Cache
+              uses: actions/cache@v1
+              with:
+                  path: node_modules
+                  key: ${{ runner.os }}-12-${{ hashFiles('**/yarn.lock') }}
+            - name: Install Deps
+              run: yarn
+            - name: Run ESLint
+              uses: icrawl/action-eslint@v1
+              env:
+                  GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+              with:
+                  job-name: ESLint
+
+    Typescript:
+        name: Typescript
+        runs-on: ubuntu-latest
+        steps:
+            - name: Checkout Project
+              uses: actions/checkout@v1
+            - name: Use Node.js 12
+              uses: actions/setup-node@v1
+              with:
+                  node-version: 12
+            - name: Restore CI Cache
+              uses: actions/cache@v1
+              with:
+                  path: node_modules
+                  key: ${{ runner.os }}-10-${{ hashFiles('**/yarn.lock') }}
+            - name: Install Deps
+              run: yarn
+            - name: Build Source
+              uses: icrawl/action-tsc@v1
+              env:
+                  GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+              with:
+                  job-name: Typescript