--- a/.prettierignore
+++ b/.prettierignore
@@ -1,1 +1,0 @@
-src/characters.js