--- a/jest.config.js
+++ b/jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+	displayName: 'ts-jest',
+	preset: 'ts-jest',
+	testEnvironment: 'node',
+	testMatch: ['<rootDir>/test/*.test.ts']
+};