--- a/tsconfig.eslint.json
+++ b/tsconfig.eslint.json
@@ -1,0 +1,8 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"module": "commonjs",
+		"baseUrl": "./"
+	},
+	"include": ["src", "test", "rollup.config.ts", "scripts"]
+}