--- a/tsconfig.json
+++ b/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "@favware/ts-config",
+	"compilerOptions": {
+		"declarationMap": true,
+		"importHelpers": false
+	},
+	"include": ["src"]
+}