--- a/typings/index.d.ts
+++ b/typings/index.d.ts
@@ -1,4 +1,0 @@
-declare module 'confusables' {
-  export function remove(str: string): string;
-  export function obfuscate(str: string): string;
-}