… .eslintignore
… .eslintrc.json
… LICENSE
… scripts/blacklistedchars
… scripts/newchars
… src/characters.js
… src/characters.ts
… src/index.js
… test/test.js
… yarn.lock