Rewriting to TS allowed us all to take a good look at the code and write
out a lot of performance boosts as well.

Some other notable changes:

- Bumped to version 1.0.0. While the rewrite is not breaking for
end-users that imported as per instructed, it is breaking in that the
whole code has been rewritten.
- Changed microbundle for rollup (which it used in the background
anyway).
  - Warnings were being thrown about incorrect usage of
rollup-plugin-node-resolve. On closer inspection it also turns out
microbundle was using outdated versions 2 core rollup plugins so yarn
was also throwing warnings
- implemented Jest as testing framework
- implemented husky to run pretty-quick
- changed txt based license for md based one because markdown is fancy
schmancy
- implemented a default export for remove, glory to ES Modules

<details><summary>Benchmarks</summary>

<h3><b><i><a href="https://hasteb.in/taxisoga.html">Based on this test suite</a></i></b></h3>

<h3><b><i><a href="https://github.com/gc/confusables/files/4049861/benchmarks.zip">Zip to download benchmarks and run them yourself</a></i></b></h3>

<h3>First Run</h3>
Old Version x 15,143 ops/sec ±0.68% (64 runs sampled)
<br/>
New Version x 212,143 ops/sec ±0.21% (66 runs sampled)

<h3>Second run</h3>
Old Version x 14,372 ops/sec ±0.45% (62 runs sampled)

<br/>
New Version x 189,943 ops/sec ±0.39% (67 runs sampled)

<h3>Third run</h3>
Old Version x 14,840 ops/sec ±0.54% (64 runs sampled)
<br/>
New Version x 211,345 ops/sec ±0.65% (66 runs sampled)

<h3>Fourth run</h3>
Old Version x 15,175 ops/sec ±0.65% (62 runs sampled)
<br/>
New Version x 210,675 ops/sec ±0.44% (66 runs sampled)

<h3>Fifth run</h3>
Old Version x 14,967 ops/sec ±0.70% (63 runs sampled)
<br/>
New Version x 208,305 ops/sec ±0.22% (67 runs sampled)

<br/>
<br/>

![source](https://user-images.githubusercontent.com/4019718/72210575-b9452080-34bd-11ea-87ed-d3afb63590e3.gif)

</details>